requirements.txt
bcrypt==4.3.0
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==45.0.7
idna==3.10
invoke==2.2.0
lxml==6.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
ncclient==0.7.0
netmiko==4.6.0
ntc_templates==8.0.0
packaging==25.0
paramiko==4.0.0
pathspec==0.12.1
platformdirs==4.4.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.0
pyserial==3.5
pytokens==0.1.10
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.1
scp==0.15.0
textfsm==2.1.0
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0

virl2-client
pyyaml
python-dotenv
